Pull out mjson.h from mjson.c to avoid duplicate symbols
mjson.c embeds the header at the very top so including .c works as well, for simple projects.
